Implement user authentication for CF Node. Design document. Authentication IDentity provider like keycloak or something like this We can add users to our CFNode deployment using the IDentity provider We can create a token per user In Studio, when creating the CF Node connection, it would also require a token, similar to what we do in Databricks
Feedback
Infinite spin in Studio when Keycloak authentication service is down.  Expected result: Studio message
CFNode provider fields in Studio are out of order
CFNode provider fields in Studio are blank when editing
Studio users can still view data after changing the token to an invalid one. Expected: Visualizations fail with an authentication error
Studio users who are importing a dashboard with a CF Node connector should be prompted to provide for clientID and token. They should not get this information in the exported CFS file.